+++ CHANGELOG.md
@@ -6,14 +6,16 @@
 
 ### 🚀 Features
 
-- Add the flag `skip_superimposed` to `segment.flattened()` which allows the user to flatten segments without flattening superimposed elements (see #664) (@roussel-ryan)
+- Add a `skip_superimposed` flag to `segment.flattened()` to allow the user to flatten segments without flattening superimposed elements (see #664) (@roussel-ryan)
 - `TransverseDeflectingCavity` now supports `linear` as a `tracking_method`. (see #678) (@cr-xu)
 
 ### 🐛 Bug fixes
 
-- Fix the following Bmad lattice conversion bugs: skips parsing lines with overlay or group definitions, wildcard element references, scientific notation, +/- before unary operators, maps type/alias fields to element metadata to reduce the number of warnings (see #663) (@roussel-ryan)
+- Fix Bmad lattice conversion issues with overlay and group definitions, wildcard element references, scientific notation in expressions, unary signs, and mapping of `type` and `alias` fields to element `metadata`. (see #663) (@roussel-ryan, @jank324)
 
 ### 🐆 Other
+
+- Update openPMD dependency to renamed package (see #684) (@jank324)
 
 ### 🌟 First Time Contributors